I've got testing up for inbox, schedule, and calendar. The easy ones... *sigh* on to tackle the challenging ones.
I also fixed some problems on schedule and calendar that were
exposed by the unit tests.

As a funny side note, calenar won't pass all it's tests until after
the 25th of april. Just the way I coded it and made the test data.
Not a big deal so I'm just going to let it roll till then.