I've fixed up emailprinting quite a bit. I would still like to flatten it before I consider it really done. There is also the way I've managed daemonization. I mean fork is good in all, but I really wish I understood why cron hates me.
I also broke the "examples" apart a bit. Mostly because the vehicle
booking and email printing utilities are utilities or use cases more
than they are examples. I'll work on some more examples, and probably
some day I'll remove these from the library. But right now they are kind
of helpful to anyone trying to learn this library and I don't want to
make a repo just for them.