Fixed issue 10 where messages were being reported as sent regardless of whether they were sent or not. This can allow you to know programatically if the message was sent and have your script respond appropriately. I also added a test case for this issue and an associated .json file.
Removed fiddle.py. it was dated. it no longer served as a good
example of how to use this library.

also changed quickmail.py. It now works to use it as part of a
unix pipeline. also it's name has been changed to pipemail.py.

added some info to setup.py. that will be totaly fleshed out one of
days!