Modified the tests classes to get them testing correctly again and updated the documentation. I need to do a lot of work on the contact and group tests. group currently doesn't even have tests. But I'll probably do that tomorrow. I'm off to gym now.
Also, I pushed off the last commit to pypi as 0.6.9. This is the
last round of this alpha. When I get the test classes writen for
group and updated for contact i'm officially in alpha 7. which is
great. It means that I am keeping up with my temporal estimations
almost to the dot. Forward into awesome!