Updated EmailPrinting, tried to get it flatter. It's still not that great of a piece of software, I would quite like to overhaul it at some point. But it is a little flatter, so progress I suppose.
Added a script called fetchFile. A proof of concept for a script
that checks an email address for emailes subjected "Fetch File" and
attempts to return a file at the path in the first line of the
email. DO NOT RUN AS ROOT! That would be a very very bad idea.