Removed superflous license from the begining of each file. The project license is all that is needed. Espectially since the two were not the same license.
Fixed a minor problem with message not working properly with
setRecipients. It was making a list in a list. Caught the problem
with my unit tests. I guess they are good for something.

Fixed Issue #13. Authentication is now expected to be a tuple in
all cases. This may develop into something more with isue #16 that
joincamp brought up. I'm exploring the use of OAuth2, but it is
quite a different world from what I'm used to. But so was many
things before I did projects like this.