Add event settings for timezones and type of event
Fix: Function description is SetAttendees while the function was named SetAttendee. Renamed the function to match the description
Add: Ability to set the timezone for the start of the meeting
Add: Ability to set the timezone for the end of the meeting
Add: Ability to specify the type of meeting (Required or not)
Add: Missing import for group, this is called by event.py to create a group object.
vent makes use of the group module on line 333 of this file. It was never imported and so that caused an error. Added group to the imports, this should relieve the issue.