Issue #34
I've turned logging off by default. This comes as a feature request from a
@sven700c but also a realization of how things normally ought run. Since
this library is meant to be used as a building block for other
applications it shouldn't have it's own independant logging system. By
turning logging off by default no log is made. But if the user decides he
wants a log, can create one and then the O365 library will attach to it.