Attempting ssl verification skipping.
I've made an alteration to the library to allow for you to choose
to ignore ssl verification. This is in response to issue #60. I've
not really tested this though. My computer is not behaving but I
want to get this change pushed. So I'll push it now and try to get
my machine to behave later.