New version with pull request #64 included.
This problem would cause problems for a number of situations, so
I feel it is important to get it out now. In the future I should
probably be a little more through before I cut releases with new
features like the ignore verify option, as that is what caused
this. Really I just need to give this library more love in general.