Fixes for Issue #78
Some quick little fixes. First there was a problem where if you
tried to get messages but had put filters on such that there were
no messages that would be returned, an error would be thrown. I
fixed that by having a simple check before going into the code
that pulls messages out of the inbox.

This came up because some one was trying my sample code exactly as
written without reading what it did. So I also changed the example
code in the README.md file to do more of what you would expect.

Lastly there was a similar bug in the fluent message class. I just
added a try catch around the fetching of the body content to keep
it from erroring out. Technically breaks the zen of python, but
it is good enough for now until a dicission is made on how exactly
to handle such a situation.