Begining experimentation into a radically different API.
I mused on github to @roycem90 about my idea of making O365 more
pythonic. I finally had a chance when I wasn't doing a lot else to
fiddle around with it and see how it would pan out. The results I
have managed so far are quite satifying. This library could be made
a lot more pythonic in a lot of ways. In terms of the code itself
I want to leverage properties more and remove the need for methods
like "getRecipient" and "setRecipient" and instead just use the
property "recipient". Apply that to all of the objects and their
data members and it should result in a pretty clear and consise
API.

The other way I want to make this a batteries included kind of
library is inspired largely by roycem90's fluent inbox and message.
Also inspiration for libraries like pandas. I've setup in this
commit fluent_inbox with a __getitem__ method that allows for
accessing the messages in the inbox super easily. I'm going to talk
with roycem and @janscas about this and see if I can get their
input on how to move forward.