Dummy tests.
I've wiped out the old content from the test files for the time
being as janscas said they would need to be rewritten. I'll
probably use the methods he had made as a skeleton in the future,
but stripping the classes like this for now will allow this branch
and pull request to checkout until I get around to writing real
tests.