Update connection.py
I have found that I couldn't assign a single value tuple without a trailing comma. If credentials is provided as `(app_id)` it ends up being turned in to a str (even if you explicitly cast it as a tuple) thus failing the tuple instance check. If you use `(app_id,)` everything works.

I added a check that will ensure it is a single value tuple by converting string to tuple (using the necessary trailing comma).