Allow a dynamic consent process using functions
This paves the way for more elegant alternatives for attaining a token using a consent URL. Example implementations would not be limited to using a web browser, a pyperclip copy, or a web browser combined with a custom redirect_uri and a temporary HTTP server on localhost.
The default and current consent process function is O365.utils.consent.consent_input_token.