Fix bug with `get_item` on `CopyOperation`
method `CopyOperation.get_item` only worked if the copied item was copied to the same location as the original item. Here we introduce the optional argument target which defaults to the parent if not given (i.e. the scenario where we copy to the same location as original item) so that the copied item can be successfully retrieved from it's ne location upon successful copy completion